feat: basic 3d mpm
- Implement 3D Material Point Method (MPM) solver (MPMSolver3D on CPU, MPMSolver3DCUDA on CUDA)
- Implement 3D Fluid-Structure Interaction (FSI) coupler (FSICoupler3D) for pressure penalty force exchange
- Purge all 3D AMR and 3D subgrid/submesh code from the C++ solver and UI, running 3D simulations strictly on uniform grids
- Add MPMDomain3D, MPMObject3D, and FSICoupler3D nodes to the frontend canvas and property editor
- Render 3D MPM particles in the WebGL/Three.js 3D viewport
- Add CPU-GPU equivalence validation test for the 3D solver

+++ backend/BlastSolver/cfd_tile.hpp
@@ -75,7 +75,7 @@ struct PrimitiveTile3D<RealType, false> {
     RealType uy[TILE_CELLS_3D];
     RealType uz[TILE_CELLS_3D];
     RealType p[TILE_CELLS_3D];
-    int floor_status[TILE_CELLS_3D];
+    uint8_t floor_status[TILE_CELLS_3D];
     RealType peak_overpressure[TILE_CELLS_3D];
     RealType running_impulse[TILE_CELLS_3D];
     RealType peak_impulse[TILE_CELLS_3D];
@@ -92,7 +92,7 @@ struct PrimitiveTile3D<RealType, true> {
     RealType alpha2[TILE_CELLS_3D];
     RealType arho1[TILE_CELLS_3D];
     RealType arho2[TILE_CELLS_3D];
-    int floor_status[TILE_CELLS_3D];
+    uint8_t floor_status[TILE_CELLS_3D];
     RealType peak_overpressure[TILE_CELLS_3D];
     RealType running_impulse[TILE_CELLS_3D];
     RealType peak_impulse[TILE_CELLS_3D];
